Deploy 3rd and 4th contracts
These are contracts which we are going to deploy on the blockchain but not index in the smart contract search engine. The reason being, we want to optimize the self discovery i.e. create a mechanism for users to test how long the self discovery process takes on a given blockchain.