Update to use esss
We now use `esss.updateStateOfContractAddress()` to push the new data into the index.
We also now use `esss.searchUsingAddress()` to get the data back out of the index.
This deliberate index and retrieval behaviour provides a better guarantee than self discovery. In order to call a public function of a smart contract, the DApp needs, and therefore has at hand, the instance address and the ABI.
So we simply use the abi and address to deliberately tell the search engine (via the new esss functions) that we have changed the data and that the search engine needs to go and reindex it now.