Simplify this demo because we will do another Eth MainNet specific demo with the new features
The web3 contract.new returns two callbacks 1) when a tx hash becomes available and 2) when the address becomes available. The new ES-SS.JS features help the smart contract search engine to deal with this on a 15 second interval blockchain. However, the SecondState Devchain will always deploy the contract and return an address within one second so this is not the arena to show the new functionality. Just keep this demo simple so that potential users of BUIDL can quickly test a demo.